requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pyjwt
"pwdlib[argon2]"
email-validator
